perf: cached os.pathconf() call in _record_path() (#1920)
### Description

The value of `os.pathconf(self._incomplete_path, "PC_NAME_MAX")` is
cached in an attribute `Persistence._max_len` -- this avoids repeatedly
querying the operating system for the same configuration information
every time `marked_incomplete()` is called on a file/job output. (Note:
this will dramatically speed up Snakemake DAG building on compute
clusters where `os.pathconf()` is slow to return and the number of files
checked for incompleteness is very large.)

### QC
<!-- Make sure that you can tick the boxes below. -->

* [X] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [X] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).