fix: Change ratelimiter dependency to throttler (#1958)
ratelimiter used outdated asyncio code that no longer works in Python
3.11. Throttler is more up to date

As part of the change, _wait_for_jobs in all ClusterExecutors have been
made async to be compatible with throttler. The Documentation on cluster
design has been updated accordingly.

Resolves #1952

### Description

<!--Add a description of your PR here-->

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

Co-authored-by: Johannes Köster <johannes.koester@uni-due.de>