feat: add  keyword to gridftp remote provide to specify the number or disable usage of multiple data stream (#1974)
### Description
- add a keyword(`streams`)to GridFTP remote provides to specify the
number or disable the usage of multiple data streams. This can be handy
when the system does not support multiple streams. The keyword can be
set in the provider and remote object(remote object is prioritized).
- Also removes some outdated info about the retry keyword for the gfal
and GridFTP provider

### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.

* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>