fix: issue #1882 WorkflowError: Metadata can't be created as it already exists (Windows) (#1971)
### Description

Fixes issue #1882: WorkflowError: Metadata can't be created as it
already exists (Windows)

The underlying cause of the problem is that `os.rename` was used to
overwrite the target file. Unfortunately, it has a different behavior
under Unix/Linux (overwrites target) and Windows (fails). `os.replace`
is an alternative to `os.rename` that simulates the Unix behavior.

The required change is extremely small (commit 4ad1ab1) and I have also
added a test case covering it (commit e12955f).

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>