fix: Correctly parse UserDicts in executors (#2016)
ResourceScopes, a subclass of UserDict (not dict), was getting parsed as
a regular iterator, not a dict, so only the keys were being passed on.

Resolves #1968

### Description

<!--Add a description of your PR here-->

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).