feat: Add token auth to GitLab/GitHub hosting providers (#1761)
### Description

This PR adds ~~an optional "token" argument~~ the ability to get a token
from environment variables to the `gitlab()` and `github()` code hosting
provider. This allows for the usage of modules stored in private
repositories.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

Closes #1301

Co-authored-by: Johannes Köster <johannes.koester@uni-due.de>
Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>