fixing bug with using flux with conda
flux can in fact have a shared filesystem, and I did some
testing today with a workflow (that I wanted to use conda)
and given the flag that there was not a shared filesystem,
I could not use conda. So I think this should be up
to the user, OR we assume that flux does have a shared
filesystem. I think having a flag that assumes a shared
fs given conda is wanted is a reasonable approach to start.
I will do more testing in the next few weeks.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>