fix: propagate attempt count from group to subjobs
Attempts was getting updated on the group job, but not getting updated
on the group job's member jobs. This prevented the member job resources
calculations from being updated, preventing the correct functioning of
resource lambdas in the group job context.

This propagates the update to the members and adds a test

Resolves #2045