fix: parsing error when last line is comment (#2054)
This occurred when a file-terminating comment was followed a keyword, as
the closing parentheses would be put on the same line as the comment and
not seen by the parser, leading to a syntax error.

Fix by adding a newline before the closing parentheses. Adds a test case
based on the original report.

Resolves #1943


### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).