fix: prevent overriding of retries when set to 0 (#2053)
The retries property of each rule was being set via the cli flag
whenever the retries directive was falsy. This was leading `retries: 0`
to be incorrectly overridden.

Fix by explicitly checking for None

Resolves #2047

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).