fix: propagate attempt count from group to subjobs (#2052)
Attempts was getting updated on the group job, but not getting updated
on the group job's member jobs. This prevented the member job resources
calculations from being updated, preventing the correct functioning of
resource lambdas in the group job context.

This propagates the update to the members and adds a test

Resolves #2045

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).