fix: skip type checks of missing dir in touch mode (#2051)
When using the --touch flag, missing files are disregarded by
dag.check_and_touch_output. This same function also checks that
directory outputs are correctly labelled with `directory()` and
vice-versa. This test was failing when the file was missing.

This patch skips the test when the file is missing.

Resolves #211


### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).