docs: Fix typo in SLURM help text (#2049)
### Description

I noticed this typo in the documentation, but could only find the text
in the actual source code. I assume something is copying help text into
the docs, but if it's hiding somewhere else, please let me know.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [ ] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).