fix: Fix handling of --jobs in no-exec state (#2029)
When running in any no-exec mode (e.g. `--dag`, `--report`), `--jobs`
was getting left as a string, leading to workflow errors.

This PR takes the opportunity to refactor the jobs and cores parsing
code, which was very convoluted. Parsing is split into three basic
possibilities: `no_exec` mode, `non_local_exec` mode, and `local_mode`.
`jobs` and `cores` are both set appropriately given the context.

A test is provided to test the new parsing function.

Resolves #1589


### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).