fix: Update list of python versions in classifiers (#2020)
Adds python versions 3.8-3.11 to classifier list

Brought to my attention by #1990: this should cause the badge to update
with the correct versions.


### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).