refactor: update versioneer from 0.18->0.28, clean-up setup.py/.cfg (#2065)
This PR does two main things:

- Update the vendored versioneer, which uses a configuration in
`pyproject.toml`, so clean-up `setup.cfg`
- A follow-up from
[#2026](https://github.com/snakemake/snakemake/pull/2026) is that most
of `setup.py` was to be cleaned-up, but this somehow missed that PR!
Keep the repeated `name="snakemake"` metadata in setup.py for now to
avoid a [GitHub bug](https://github.com/orgs/community/discussions/6456)

<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [ ] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).