refactor: move static metadata (again) to setup.cfg (#2067)
### Description

This moves the static project metadata (again) to setup.cfg, which has
had long-standing support with
[setuptools](https://setuptools.pypa.io/en/latest/userguide/declarative_config.html).

This resolves an issue with limitations of project metadata in
pyproject.toml by release-please.

### QC

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [ ] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).