fix: Postprocess job groups in toposorted order for correct touch times (#2073)
### Description

If I understand the mechanics correctly, jobs in job groups executed on
clusters get "postprocessed", and as part of this their output files are
touched. However, the order in which this happened was not necessarily a
topological order, meaning that output files of dependent rules could be
touched in the "wrong" order. See
https://github.com/snakemake/snakemake/issues/789

This simple change just ensures that jobs are post-processed in
topological order, which should solve the above issue.

If desirable, I could also writing a test for the above behaviour to
catch potential future issues.

### QC
* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

Co-authored-by: Johannes Köster <johannes.koester@uni-due.de>