fix: assume shared filesystem by default when running with --flux (#2075)
Signed-off-by: vsoch <vsoch@users.noreply.github.com>

### Description

The `--use-conda` flag does not work with Flux! This is my attempt to
reproduce what I see locally. I'll mark this as ready when I see it
reproduced here.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [ ] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

---------

Signed-off-by: vsoch <vsoch@users.noreply.github.com>
Co-authored-by: vsoch <vsoch@users.noreply.github.com>
Co-authored-by: Johannes Köster <johannes.koester@uni-due.de>
Co-authored-by: Johannes Koester <johannes.koester@uk-essen.de>