feat: add cleanup containers option (#2088)
### Description

Closes #2078. I've created an isolated test case, and have also tested
it in the wild. Couldn't figure out how to make a test for a proper
pipeline run though. Hope the test case I've added is sufficient though.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).