fix: batch collect jobs for scancel (#2114)
### Description

When using --slurm, exiting snakemake by SIGINT or ctrl-c helpfully
cancels spawned jobs. However, this is quite unreliable as it often
hangs for a few minutes and exits without canceling the jobs in the end
(without indicating that it failed either).

Slurm documentation on
[scancel](https://slurm.schedmd.com/scancel.html#SECTION_PERFORMANCE)
notes that a large number of scancel calls at the same time may result
in denial of service.

Snakemake runs scancel on each job
[individually](https://github.com/snakemake/snakemake/blob/main/snakemake/executors/slurm/slurm_submit.py#L136).
Instead, job ids should be collected and cancelled all at once.

fixes #2113

### QC
* [X] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [X] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>