fix: better job status queries for slurm executor  (#2136)
### Description

This PR changes two things:

1. Previously, checking `active_jobs` was only initiated after sleeping
for `1/max_status_checks_per_second` seconds, which meant a minimum
sleeping time of 30 seconds. This sleeping interval is now set to a
fixed 30 seconds. However, one idea could also be to turn this into a
new variable for the generic executor, and expose it to the user via a
command-line argument -- it is also manually set for other executors
(e.g. the google executor), and might require fine-tuning for specific
workflows and/or clusters.
2. Previously, a `max_status_checks_per_second > 1` specified by the
user was set down to `max_status_checks_per_second = 0.03` (no clue why
there were different values in the comparison and the final setting) and
the default value was also `0.03`. This means that only one job was
checked for its status every 30 seconds. For workflows with hundreds or
thousands of short-running jobs, this meant that it could easily take
hours for finished jobs to be registered as `COMPLETED` and until that
point, the respective resources were not available for scheduling new
jobs. The new default and maximum of `max_status_checks_per_second = 2`
considerably speeds this up, but stays well below the number of status
checks that would be possible on the Slurm cluster I tested this on
(which would be almost 20 for `sacct` and more than 50 for `scontrol`).
Here, it would be useful if users on other Slurm clusters could check,
whether `scontrol` is also always quicker for them, compared to `sacct`.
If so, it would make sense to make `scontrol` the (cheaper) default
(which also works for jobs that are not yet started), and `sacct` the
fall-back (which also works for jobs that have already finished a while
ago and might not be found by `scontrol` any more).

### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

---------

Co-authored-by: David Laehnemann <david.laehnemann@helmholtz-hzi.de>