perf: Gfal2 remote provider using gfal2-python instead of  gfal2-utils. (#2128)
### Description

Added support of gfal2-python library instead of gfal2-utils.
Time(wallclock) of a dry run with 100 files on a gridftp server was
reduced by a 7 to 11 fold (from about a minute to 6 seconds!), while the
system and user time was also significantly less. This speed-up is
caused because no shell->python->gfal2 objects have to be generated for
every file (gfal2-utils are gfal2-python commands wrapped with a command
line interface).

If gfal2-python is not installed, it will fall back to gfal2-utils. This
makes the code unclear, but this will not break the current installation
when upgrading. A deprecation warning is given if gfal2-python is not
installed.

### QC


* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
Test case was not written because of lack of generic available server
and adding dependencies
* [X] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behaviour or functionalities of Snakemake).
Should not change behaviour

---------

Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>