fix: 2142 log directive in default target rule (#2191)
### Description

When the default target rule (e.g `all`) contains a `log` directive,
snakemake deems workflow complete and skips execution. See #2142

This PR fixes this bug by adding a `include_logfiles` flag to
`job.has_product()` method, similar to one used in `job.products()`. In
fact, this flag is simply passed on to `job.products()` internally.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).