fix: Use `job.rule.name` attribute to fill rule field in summary (#2217)
### Description

This PR fixes #2216, by using the job rule name as derived in the dag.

I am not entirely sure what the persistence class does here, but I guess
it somehow caches the jobname in the .snakemake directory when the rule
is run. This probably leads to incorrect summaries before the rules have
been cached, which does not seem like the intended behavior of the
summary option.

### QC

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).