docs: update misc/vim/Readme with info for packer.nvim (#2095)
### Description

`packer.nvim` is a popular lua-based packet manager for nvim, this info
might help a lot of people.
The added line can go into the lua config file for nvim.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case. -> no code changes
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake). -> Docs link
to updated Readme

Co-authored-by: Henning Timm <henning.timm@tu-dortmund.de>