refactor: format()  and %s replaced by f-strings (#2249)
### Description

Python support f-string since python 3.6. Formatting strings with
f-string has an higher readablity due to easier and more compressed
syntax and is a tiny bit faster. This pull request rewrite format() and
%s syntax to f-strings. However, there are still places were format is
used because it is not straightforward to convert them without risk of
breaking code. Most work is done with [ruff](https://beta.ruff.rs/docs/)

`ruff check --select UP032,UP031,F541  . `

This pull request touches a lot of lines, so this might break other work
in progress.


### QC

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).