refactor: removed unused import (#2265)
### Description

Deleted unused import statements to clean the list of imports a bit. I
could not test external packages/tools that use the import in
snakemake(I can not test what I do not know) , so a word of caution
here.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).