fix: addressing #2197 by allowing 256 character account names in slurm (#2198)
### Description

This changes the format string on the calls to 'sacct' and 'sacctmgr' to
allow for long slurm account names. This fix seems to work on our slurm
cluster.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).