refactor:  more format() and %s replaced by f-strings (#2313)
### Description
[
Python supports f-string since Python 3.6. Formatting strings with
f-string has a higher readability due to easier and more compressed
syntax and is a tiny bit faster. This pull request rewrite format() and
%s syntax to f-strings. However, there are still places where format is
used because converting them without the risk of
breaking code is not straightforward. Most work is done with
[ruff](https://beta.ruff.rs/docs/)

`ruff check --select UP032,UP031,F541  . `

This pull request touches a lot of lines, so this might break other work
in progress.

(continuation of https://github.com/snakemake/snakemake/pull/2249)
### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).