fix: conda env inside script (#1812)
### Description

Small fix which enables to run conda envs inside `script:`. Because path
to conda was not absolute, python executable also wasn't defined
correctly. This fixes the issue described in
[#1811](https://github.com/snakemake/snakemake/issues/1811).

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

---------

Co-authored-by: David Laehnemann <david.laehnemann@hhu.de>
Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>
Co-authored-by: Johannes Köster <johannes.koester@uni-due.de>