fix: only last daughter process accounted in benchmark cpu time (#2278)
Solves #2147  by keeping track of CPU times in a dictionary. 

### Description

The CPU time was computed at each call of
``BenchmarkTimer:_update_record`` by looping over the current
``chain((self.main,), self.main.children(recursive=True))`` and summing
the CPU times.

This approach only account for the total CPU times of processes that are
still alive when the last ``BenchmarkTimer:_update_record`` is done, so
when 2 or more processes are started sequentially, the CPU times value
only capture the running time of the last process.

To correct this bug, I propose to replace
``BenchmarkTimer:processed_procs`` by a dictionary and keep track of the
total CPU times of all processes spawn by the main current process. It
know gets the proper result when using the example provided in #2147.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

---------

Co-authored-by: Valentin Pestel <vpestel@km3net.de>
Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>