fix: handle non-PEP440 versions of apptainer/singulariy (#2337)
### Description
Apptainer and singularity use non-PEP440 version numbers like
`1.1.8-1.el8` and `3.10.4-dirty`. packaging.version can not handle these
version numbers. To mitigate the issue, a character is stripped away at
the end of the string until it is parsable by packaging.version.

Should fix #2319  and #2334

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

---------

Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>
Co-authored-by: Johannes Köster <johannes.koester@uni-due.de>