docs: clarify the channel priority in environment definition deployment.rst (#2352)
### Description

In Snakemake documentation, the priority of channels is not known to the
user in environment definitions. I attempted to make this information
explicit in case people with little knowledge of conda have issues with
it.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [ ] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).