fix: require python >=3.7 again (the python 3.9 dependency was unnecessary) (#2372)
### Description

The respective bug was introduced with PR #2341, so it exists starting
with version `7.30.2`. Also see issue #2369 and PR #2370 for further
context.

This code from PR #2341 seems to get run every time a conda environment
is used:


https://github.com/snakemake/snakemake/blob/13b5d4d11d7475f942531c51b6013962a88a72b7/snakemake/__init__.py#L8-L9

Thus, any conda environment in a rule with an older python version
included, will trigger a:
```
  File "/opt/conda/envs/snakemake/lib/python3.11/site-packages/snakemake/__init__.py", line 9, in <module>
    raise ValueError("Snakemake requires at least Python 3.9.")
ValueError: Snakemake requires at least Python 3.9.
```

This would preclude snakemake from running any software that needs an
older python version, so this is clearly an unwanted side-effect. But I
am not sure how to fix this.

Currently, this PR is only a minimal example test to trigger this
behaviour in the CI tests. Hopefully, this will make debugging it
quicker for @johanneskoester (or someone else?).

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation (`docs/`) is updated to reflect the changes or
this is not necessary (e.g. if the change does neither modify the
language nor the behavior or functionalities of Snakemake).

---------

Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>
Co-authored-by: Johannes Köster <johannes.koester@uni-due.de>